requirements.txt
google-api-python-client>=2.111.0
python-dotenv==1.0.0
anthropic>=0.39.0
requests>=2.31.0
